feat: login improvements
CSRF now persists acorss all calls and CORS allows for scripting access
Login now works in tandem with the CSRF token, so POSTS cannot be made with
having a CSRF token. The LoginResponse type has now also been defined for more
type safety between the front and backend. Will be making more of those as I
have the time to do so. Thought about possibly creating an
`@unteris/shared/sdk` to be able to just do something like `sdk.logIn(body)`
It's definitely a thought worth entertaining.
`secure` is now only added to the cookie if in production. This probably
wasn't an issue, but I figured it wouldn't hurt to set anyways.

Error handling on the login and registration still needs to be taken care of
and the CSRF guard should be moved to be global rather than bound per route
to ensure that **ALL** non `GET` requests come from the expected location