feat: use the email service to send verification emails
When a user now signs in for the first time, a verificaiton email will
be sent to the user wtih a link to verify their email address. Once they
verify it, the verified flag on the user's record will be switched to
true. This so far will not really have any effect, but it could help
with account clean up in the future.