feat: create new postgres project
This is generally an empty project in Nx's eyes, but it does exist for
the sole purpose of allowing me to only push up a new postgres image
whenever I modify the dockerfile for the unteris postgres image.

The reason for the image is to make use of the [ulid][ulid] extension
that actually has a true ulid type in postgres and can be translated
into timestamps by casting. They're also way faster, which is a big
bonus.

The dockerfiles have also been split apart again, which, while creating
some WET code in terms of the dockerfiles, it does allow for a better
separation of concerns of each file, and can probably be used to better
improve each file as necessary. The local docker compose file also now
uses the local postgres dockerfile instead of the common image, again
for the ulid implementation.

The docker build executor now takes in these dockerfiles by matching to
the project root and builds each image. Seems to be faster as well, but
only time will tell when it comes toactually publishing.

Speaking of which, I should start running the package non-publish
command on CI runs to ensure when I merge to main it will still work as
expected.

[ulid]: https://github.com/pksunkara/pgx_ulid/