fix: create the migrations for the conversion of tables to ulid
Every text id in the database will need to be converted to the new
`ulid` type once the extensiion allows for joins. This conversion set
will update all of the types, and set up the defaults to be the new
`gen_ulid` method that the extension provides.

Once everything is in working order, I'll create another migration that
drops the original `ulid()` function I was working with.