feat: email verificaiton page
Wording probably needs some updating, but it'll work for now.
The User page has also been moved to a separate route rather than a part
of login, and the profile button could probably be made to account
for that along with having the login page automatically redirect in
the case of already being signed in.
Lastly, a new fetch helper has been added `getFetch`, for simple get
requests. This is to be used with async atoms from jotai.