feat: create image processing service
The image processing microservice now has an event it listens to through
rabbitmq. It takes in an image id and will processes the image to have
three sizes saved on the disk, small (half size), medium (same size), and
large (triple the size), saved as webp format for the sake of more updated
image format types. Then the image record is updated with the new urls.