fix: strict same site cookies non-json logging and dev docker compose
As I look at the docker logs directly, and not through something like
grafana or prmetheus, it's better to keep them in the stream string format.

Cookies should now be set to SameSite=Strict to allow them to work properly
in a production environment (hopefully).

There's now a docker-compose.dev.yml that immitates the docker-compose.prod.yml
to allow for a "full test" of the system